Separate SerialBatch and ParallelBatch.
